fix: native password and kk pin modal migrated
