+++ src/context/WalletProvider/KeepKey/hooks/useKeepKeyEventHandler.ts
@@ -2,36 +2,40 @@ import { Event, Events } from '@shapeshiftoss/hdwallet-core'
 import { Dispatch, useEffect } from 'react'
 import { ActionTypes, WalletActions } from 'context/WalletProvider/actions'
 import { InitialState } from 'context/WalletProvider/WalletProvider'
-import { useModal } from 'hooks/useModal/useModal'
 
 import { FailureType, MessageType } from '../KeepKeyTypes'
 
-type KeyringState = Pick<InitialState, 'keyring' | 'walletInfo'>
+type KeyringState = Pick<InitialState, 'keyring' | 'walletInfo' | 'modal'>
 
 export const useKeepKeyEventHandler = (state: KeyringState, dispatch: Dispatch<ActionTypes>) => {
-  const { keepkeyPin, keepkeyPassphrase } = useModal()
-  const { keyring } = state
+  const { keyring, modal } = state
 
   useEffect(() => {
     const handleEvent = (e: [deviceId: string, message: Event]) => {
+      const deviceId = e[0]
       switch (e[1].message_enum) {
         case MessageType.PASSPHRASEREQUEST:
-          if (!keepkeyPassphrase.isOpen) {
-            keepkeyPassphrase.open({ deviceId: e[0] })
+          if (!modal) {
+            // TODO: migrate keepkey passphrase modal
+            // dispatch({ type: WalletActions.OPEN_KEEPKEY_PASSPHRASE, payload: { deviceId } })
           }
           break
         // ACK just means we sent it, doesn't mean it was successful
         case MessageType.PASSPHRASEACK:
-          if (keepkeyPassphrase.isOpen) keepkeyPassphrase.close()
+          if (modal) dispatch({ type: WalletActions.SET_WALLET_MODAL, payload: false })
           break
         case MessageType.PINMATRIXREQUEST:
-          if (!keepkeyPin.isOpen) {
-            keepkeyPin.open({ deviceId: e[0], pinRequestType: e[1].message?.type })
-          }
+          dispatch({
+            type: WalletActions.OPEN_KEEPKEY_PIN,
+            payload: {
+              deviceId,
+              pinRequestType: e[1].message?.type
+            }
+          })
           break
         // ACK just means we sent it, doesn't mean it was successful
         case MessageType.PINMATRIXACK:
-          if (keepkeyPin.isOpen) keepkeyPin.close()
+          if (modal) dispatch({ type: WalletActions.SET_WALLET_MODAL, payload: false })
           break
         // @TODO: What do we want to do with these events?
         case MessageType.FAILURE:
@@ -110,5 +114,5 @@ export const useKeepKeyEventHandler = (state: KeyringState, dispatch: Dispatch<A
       keyring.off(['*', '*', Events.CONNECT], handleConnect)
       keyring.off(['*', '*', Events.DISCONNECT], handleDisconnect)
     }
-  }, [dispatch, keepkeyPassphrase, keepkeyPin, keyring, state.walletInfo])
+  }, [dispatch, keyring, modal, state.walletInfo])
 }
